fix assets not importing
After some tinkering, I finally realized that the apostrophe in the name "don't-rotate.gif" was causing the error with vitepress! It put a single quote inside a single-quoted string...

// code in the internal error
import _imports_2 from '/img/don't-rotate.gif'

vitepress complained that it was invalid JS syntax because, well, it was! Yes, this is probably an error that shouldn't happen in the first place, but it's amazing how little things like that can mess it all up.

For now, the apostrophe has been removed from the file name, and it now properly loads. Additionally, all assets have been moved to the public folder for easier organization.